interoplib: NativeCrc32 не полагается на знаковое переполнение при проверке границ
offset + count складывались signed int и лишь потом кастовались в uint32:
у двух больших положительных это переполнение, то есть UB, и проверку
компилятор вправе выкинуть целиком. Кастуем каждый операнд — как это уже
сделано в соседнем FrameDecoder, где та же ловушка разобрана в комментарии.

+++ InteropAssemblies/interoplib/interoplib_interoplib_Utilities.cpp
@@ -1,12 +1,7 @@
 //-----------------------------------------------------------------------------
 //
-//                   ** WARNING! **
-//    This file was generated automatically by a tool.
-//    Re-running the tool will overwrite this file.
-//    You should copy this file to a custom location
-//    before adding any customization in the copy to
-//    prevent loss of your changes when the tool is
-//    re-run.
+// LEDTREES: мелкие нативные утилиты (реализация пишется руками — генерируются
+// только interoplib.cpp/.h и *_mshl.cpp, их копируют из Stubs).
 //
 //-----------------------------------------------------------------------------
 
@@ -27,19 +22,22 @@ void Utilities::NativeGetBaseMac( CLR_RT_TypedArray_UINT8 param0, HRESULT &hr )
     memcpy((void*)param0.GetBuffer(), baseMac, 6);
 }
 
-// LEDTREES: инкрементальный zlib-совместимый CRC32 через ROM-функцию.
+// Инкрементальный zlib-совместимый CRC32 через ROM-функцию: табличный цикл по байту
+// на nanoCLR считает CRC мегабайтного .ltfw десятки секунд, ROM-функция — миллисекунды.
+//
 // Семантика esp_rom_crc32_le совпадает с zlib.crc32: состояние между вызовами —
 // финализированное значение (инверсия на входе/выходе внутри ROM-реализации).
-// Managed-версия считала табличный цикл по байту интерпретатором — на мегабайтном
-// .ltfw это десятки секунд на каждый OTA.
 unsigned int Utilities::NativeCrc32( unsigned int param0, CLR_RT_TypedArray_UINT8 param1, signed int param2, signed int param3, HRESULT &hr )
 {
     const uint8_t *data = (const uint8_t *)param1.GetBuffer();
     signed int offset = param2;
     signed int count = param3;
 
-    // валидация границ managed-массива: не читаем мимо
-    if (data == NULL || offset < 0 || count < 0 || (uint32_t)(offset + count) > param1.GetSize())
+    // Границы managed-массива проверяем сами: bounds-check CLR в нативном коде не
+    // работает. offset/count складываем уже беззнаковыми — у signed int сумма двух
+    // больших положительных переполняется (UB), и проверку компилятор вправе выкинуть.
+    if (data == NULL || offset < 0 || count < 0 ||
+        (uint32_t)offset + (uint32_t)count > param1.GetSize())
     {
         hr = CLR_E_INVALID_PARAMETER;
         return param0;
